requirements.txt
Flask
edge-tts
python-docx
